requirements.txt
Flask==2.2.3
flask-restx==1.1.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
marshmallow==3.19.0
PyJWT==2.6.0

pytest~=7.2.1
SQLAlchemy~=1.4.26
Werkzeug~=2.2.3